Update angular-google-maps.js
Replacing controlDiv.contents() with controlDiv.children() in Control.prototype.link. jqLite appears to have a bug in that its .append() method doesn't behave exactly like jQuery's, resulting in an error when using the jQuery and the mapControl directive and when there's leading whitespace in the ng-template script tag (which is probably always the case).